fix: switched to app router due to route resolve issues
