+++ frontend/.env-example
@@ -1,3 +1,4 @@
+NEXT_PUBLIC_APP_NAME=YrkesutbildningMitt
 NEXT_PUBLIC_API_URL=http://localhost:3001/api
 DOMAIN_NAME=localhost
 NEXT_PUBLIC_BASE_URL=http://localhost:3000
