+++ frontend/package.json
@@ -5,15 +5,16 @@
   "scripts": {
     "dev": "next dev",
     "build": "next build",
-    "postbuild": "next-sitemap",
     "start": "next start",
     "lint": "next lint --no-cache",
     "analyze": "cross-env ANALYZE=true next build",
     "analyze:server": "cross-env BUNDLE_ANALYZE=server next build",
     "analyze:browser": "cross-env BUNDLE_ANALYZE=browser next build",
     "cypress": "cypress open",
-    "test": "jest --watch",
-    "test:coverage": "jest --watchAll --coverage"
+    "cypress:headless": "cypress run",
+    "jest": "jest --watch",
+    "jest:coverage": "jest --coverage",
+    "test:coverage": "yarn cypress:headless"
   },
   "dependencies": {
     "@emotion/react": "^11.11.4",
@@ -36,7 +37,6 @@
     "lodash": "^4.17.21",
     "lucide-react": "^0.471.1",
     "next": "14.2.23",
-    "next-sitemap": "^4.2.3",
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "react-hook-form": "^7.54.2",
@@ -51,13 +51,17 @@
   },
   "devDependencies": {
     "@babel/plugin-transform-private-methods": "^7.25.9",
-    "@babel/preset-env": "^7.26.0",
+    "@babel/preset-env": "^7.26.9",
     "@babel/preset-react": "^7.26.3",
+    "@babel/preset-typescript": "^7.26.0",
+    "@cypress/code-coverage": "^3.13.12",
     "@eslint/js": "^8.57.1",
+    "@istanbuljs/nyc-config-typescript": "^1.0.2",
     "@tailwindcss/container-queries": "^0.1.1",
+    "@testing-library/dom": "^10.4.0",
     "@testing-library/jest-dom": "^6.6.3",
-    "@testing-library/react": "16.2.0",
-    "@testing-library/user-event": "^14.6.0",
+    "@testing-library/react": "^16.2.0",
+    "@testing-library/user-event": "^14.6.1",
     "@types/bcryptjs": "^2.4.6",
     "@types/jest": "^29.5.14",
     "@types/lodash": "^4.17.15",
@@ -67,15 +71,16 @@
     "@types/react-sticky-el": "^2.1.0",
     "@types/sanitize-html": "^2.13.0",
     "@types/underscore.string": "^0.0.41",
-    "@typescript-eslint/eslint-plugin": "^8.20.0",
+    "@typescript-eslint/eslint-plugin": "^8.25.0",
     "autoprefixer": "^10.4.20",
     "babel-jest": "^29.7.0",
+    "concurrently": "^9.1.2",
     "cross-env": "^7.0.3",
-    "cypress": "^14.0.0",
+    "cypress": "^14.1.0",
     "dotenv": "^16.4.7",
     "envalid": "^8.0.0",
     "eslint": "^8.57.1",
-    "eslint-config-next": "14.2.23",
+    "eslint-config-next": "14.2.24",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-jest": "^28.11.0",
     "eslint-plugin-react-refresh": "^0.4.18",
@@ -91,10 +96,9 @@
     "sass": "^1.84.0",
     "tailwindcss": "^3.4.17",
     "typescript": "5.7.3",
-    "typescript-eslint": "^8.20.0"
+    "typescript-eslint": "^8.25.0"
   },
   "browserslist": [
-    "defaults and fully supports es6-module",
-    "maintained node versions"
+    "defaults and fully supports es6-module"
   ]
 }
