+++ frontend/src/app/layout.tsx
@@ -1,13 +1,14 @@
+'use client';
+
 import LoginGuard from '@components/login-guard/login-guard';
+import { AppWrapper } from '@contexts/app-context/app-wrapper';
 import { ColorSchemeMode, GuiProvider, defaultTheme } from '@sk-web-gui/react';
 import '@styles/tailwind.scss';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import dayjs from 'dayjs';
 import 'dayjs/locale/sv';
 import updateLocale from 'dayjs/plugin/updateLocale';
 import utc from 'dayjs/plugin/utc';
-import type { AppProps /*, AppContext */ } from 'next/app';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { AppWrapper } from '@contexts/app-context/app-wrapper';
 
 // Create a client
 const queryClient = new QueryClient();
@@ -33,18 +34,27 @@ dayjs.updateLocale('sv', {
   monthsShort: ['Jan', 'Feb', 'Mar', 'Apr', 'Maj', 'Jun', 'Jul', 'Aug', 'Sep', 'Okt', 'Nov', 'Dec'],
 });
 
-function MyApp({ Component, pageProps }: AppProps) {
+function MyApp(props: { children: React.ReactNode }) {
   return (
     <GuiProvider theme={defaultTheme} colorScheme={ColorSchemeMode.Light}>
       <QueryClientProvider client={queryClient}>
         <AppWrapper>
-          <LoginGuard>
-            <Component {...pageProps} />
-          </LoginGuard>
+          <LoginGuard>{props.children}</LoginGuard>
         </AppWrapper>
       </QueryClientProvider>
     </GuiProvider>
   );
 }
 
-export default MyApp;
+export function AppLayout(props: { children: React.ReactNode }) {
+  return (
+    <html lang="sv">
+      <head />
+      <body>
+        <MyApp>{props.children}</MyApp>;
+      </body>
+    </html>
+  );
+}
+
+export default AppLayout;
