+++ frontend/src/app/logout/page.tsx
@@ -1,0 +1,5 @@
+import { Logout } from '@layouts/page-layout/logout-page.component';
+
+export default function Page() {
+  return <Logout />;
+}
