+++ frontend/src/components/header/big-drop-header.component.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { usePlaceholderImg } from '@utils/use-placeholder-image.hook';
 import Image from 'next/image';
 import React from 'react';
