+++ frontend/src/components/saved-interests/saved-interests-form.component.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import Button from '@components/button/button.component';
 import {
   categoryFilterPlaceholder,
