+++ frontend/src/layouts/page-layout/utbildningar-efterfragade-efterfragad.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import ContentBlock from '@components/block/content-block.component';
 import Breadcrumbs from '@components/breadcrumbs/breadcrumbs.component';
 import Button from '@components/button/button.component';
@@ -17,10 +19,7 @@ export const Efterfragade: React.FC<EfterfragadeProps> = ({ layoutData, pageData
   if (!educationData) return <></>;
 
   return (
-    <DefaultLayout
-      title={`Yrkesutbildning - Efterfrågade av arbetsgivare - ${educationData.title}`}
-      layoutData={layoutData}
-    >
+    <DefaultLayout layoutData={layoutData}>
       <ContentBlock>
         <BigDropHeader
           imageSrc={pageData?.imgSrc}
