+++ frontend/src/utils/app-name.ts
@@ -1,0 +1,1 @@
+export const appName = () => process.env.NEXT_PUBLIC_APP_NAME ?? 'APP_NAME';
